Fix image preview modal & import bib.js changes from studio_webapp
Corrected the class names for the image preview modal from the image library to be the same size on all desktop screens & full screen on all mobile devices.

Copied all the bib.js changes from studio_webapp as the jplist funcitonality cannot be done on the front-end stnadalone framework. It requireres calls to the server. This should be the only file that is worked studio_webapp => vle_front-end.